refactor(ui): consolidate create/assign/date primitives under app/core/ui (#293)
Move rhythm_task_create_bar, rhythm_assignee_field, and rhythm_date_button
from shared/widgets into app/core/ui alongside the existing
rhythm_task_create_dialog so the design-system primitives live in one
home. Export all four from rhythm_ui.dart and drop now-redundant explicit
imports across tasks, projects, weekly planner, dashboard, and the
inspector primitive.

Closes #293
Refs #244
