+++ apps/desktop_flutter/lib/app/core/ui/rhythm_ui.dart
@@ -1,11 +1,13 @@
+export 'focus_business_widgets.dart';
+export 'rhythm_assignee_field.dart';
 export 'rhythm_badge.dart';
 export 'rhythm_button.dart';
 export 'rhythm_compact_row.dart';
+export 'rhythm_date_button.dart';
 export 'rhythm_detail_pane.dart';
-export 'rhythm_disclosure.dart';
 export 'rhythm_dialog.dart';
+export 'rhythm_disclosure.dart';
 export 'rhythm_empty_state.dart';
-export 'focus_business_widgets.dart';
 export 'rhythm_filter_bar.dart';
 export 'rhythm_inspector.dart';
 export 'rhythm_menu_button.dart';
@@ -16,5 +18,7 @@ export 'rhythm_search_field.dart';
 export 'rhythm_section_header.dart';
 export 'rhythm_segmented_control.dart';
 export 'rhythm_surface.dart';
+export 'rhythm_task_create_bar.dart';
+export 'rhythm_task_create_dialog.dart';
 export 'rhythm_toolbar.dart';
 export 'tokens/rhythm_theme.dart';
